Attempt 001: Fixing parent problem attempt
Debbuging revealed that when currNode got a new value (=) all the parent node changed. Probably because they were pointing to that memory space.
Now I made it allocate memory (new) each time it runs throgth the loop and now it works the first time.
It generates a path but if you try it crashes because it accesses to invalid memory. It may or may not be realated, I'm looking into it.