Fixed an error that caused it to get caught in an infinite loop
Changing the lists (open and closed) to be local variables was the solution.
They weren't reseted at any moment so when it visited nodes that were already in those lists it updated nodes or not (if they were at a shorter distance) making PathNode::parent not point to the correct node.